# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Telegram Group Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-code:
    post:
      operationId: sendAuthCode
      tags: [auth]
      summary: Send Telegram auth code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendCodeBody"
      responses:
        "200":
          description: Code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/verify-code:
    post:
      operationId: verifyAuthCode
      tags: [auth]
      summary: Verify auth code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCodeBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/verify-password:
    post:
      operationId: verifyAuthPassword
      tags: [auth]
      summary: Verify 2FA password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPasswordBody"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /search/groups:
    get:
      operationId: searchGroups
      tags: [search]
      summary: Search Telegram groups
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
        - name: minMembers
          in: query
          required: false
          schema:
            type: integer
        - name: maxMembers
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: groupType
          in: query
          required: false
          schema:
            type: string
            enum: [all, group, channel, supergroup]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupSearchResults"
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List saved groups
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, saved, joined, pending]
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupList"
    post:
      operationId: saveGroup
      tags: [groups]
      summary: Save a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveGroupBody"
      responses:
        "201":
          description: Group saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
...
  /groups/join:
    post:
      operationId: joinGroups
      tags: [groups]
      summary: Auto-join selected groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupsBody"
      responses:
        "200":
          description: Join job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResult"
...
components:
  schemas:
...
    AuthResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        requiresPassword:
          type: boolean
          nullable: true
        authenticated:
          type: boolean
          nullable: true
      required: [success, message]
    SendCodeBody:
      type: object
      properties:
        phone:
          type: string
      required: [phone]
    VerifyCodeBody:
      type: object
      properties:
        phone:
          type: string
        code:
          type: string
        phoneCodeHash:
          type: string
      required: [phone, code, phoneCodeHash]
    VerifyPasswordBody:
      type: object
      properties:
        password:
          type: string
      required: [password]
    TelegramGroup:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        username:
          type: string
          nullable: true
        membersCount:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        type:
          type: string
          enum: [group, channel, supergroup]
        isPublic:
          type: boolean
        inviteLink:
          type: string
          nullable: true
      required: [id, title, type, isPublic]
    GroupSearchResults:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/TelegramGroup"
        total:
          type: integer
      required: [results, total]
    Group:
      type: object
      properties:
        id:
          type: integer
        telegramId:
          type: string
        title:
          type: string
        username:
          type: string
          nullable: true
        membersCount:
          type: integer
          nullable: true
        type:
          type: string
        status:
          type: string
          enum: [saved, joined, pending, failed]
        createdAt:
          type: string
      required: [id, telegramId, title, type, status, createdAt]
    GroupList:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        total:
          type: integer
      required: [groups, total]
    SaveGroupBody:
      type: object
      properties:
        telegramId:
          type: string
        title:
          type: string
        username:
          type: string
          nullable: true
        membersCount:
          type: integer
          nullable: true
        type:
          type: string
      required: [telegramId, title, type]
    JoinGroupsBody:
      type: object
      properties:
        groupIds:
          type: array
          items:
            type: integer
        delaySeconds:
          type: integer
          default: 5
      required: [groupIds]
...
    JobResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        jobIds:
          type: array
          items:
            type: integer
          nullable: true
      required: [success, message]
...